🎲 Implemented the PRNG module.
This module currently consists of four classes:
1- TRandomGenerator: an interface for all PRNGs
2- JavaSplittableRandom: a wrapper class for java.util.SplittableRandom to work with Tephrium's interface
3- RNGUtils: containing RNG related utility functions such as mixers, initial seed generators etc
4- SplitMix64Random: the main focus of this commit, an implementation of splitmix64

HUGE THANKS TO TOMMY ETTINGER (https://github.com/tommyettinger/) FOR HELPING ME WITH THIS COMMIT.
He taught me a lot of bitwise tricks and explained a lot of RNG related algorithms to me.
This commit would never be possible without him.
He didn't write any code or contributed any copyrightable things to the project
but he personally gave me a lot of articles to read and helped for full 3 days!

So huge thanks to him for helping me!